d-ai-app-patterns/scalable_chat/src/front_service/pyproject.toml
[project]
name = "scalable-chat-front"
version = "0.1.0"
description = "Front-end service for scalable chat using SSE and Azure Service Bus"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.112.0",
    "azure-identity==1.23.0",
    "azure-servicebus==7.14.2",
    "python-dotenv==1.1.0",
    "uvicorn[standard]==0.29.0",
    "aiohttp>=3.11.18",
    "opentelemetry-sdk>=1.33.1",
    "opentelemetry-instrumentation-fastapi>=0.54b1",
    "azure-monitor-opentelemetry>=1.6.7",
]
